Move getHigherType() to TypeInfo
